requirements.txt
Flask==0.12.2
flask_bootstrap==3.3.7.1
flask_wtf==0.14.2
mysql_connector_repackaged==0.3.1
WTForms==2.2.1
